requirements.txt
python-dateutil~=2.8.2
SQLAlchemy~=1.3.10
SQLAlchemy-Utils~=0.38.2
uvicorn~=0.17.6
fastapi~=0.75.0
numpy~=1.22.2
pandas~=1.4.1
psycopg2~=2.9.3